# openapi/user-api.yaml
openapi: 3.0.3
info:
  title: User API
  version: 1.0.0
  description: CRUD API for managing users

servers:
  - url: http://localhost:8080/

paths:
  /users:
    get:
      summary: Get all users (optionally filtered by role)
      operationId: getAllUsers
      tags:
        - User
      parameters:
        - in: query
          name: role
          required: false
          schema:
            type: string
            enum: [ROLE_ADMIN, ROLE_USER, ROLE_WAREHOUSE, ROLE_MANAGER]
          description: Optional role to filter users
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
    post:
      summary: Create a new user
      operationId: createUser
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'


  /users/{id}:
    get:
      summary: Get a user by id
      operationId: getUserById
      tags:
        - User
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

    patch:
      summary: Update user fields
      operationId: updateUser
      tags:
        - User
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPatchRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

    delete:
      summary: Delete a user
      operationId: deleteUser
      tags:
        - User
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User deleted successfully

  /users/usernames:
    get:
      summary: Get all usernames
      operationId: getAllUsernames
      tags:
        - User
      responses:
        '200':
          description: List of usernames
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
      security:
        - BearerAuth: [ ]

  /users/username/{username}:
    get:
      summary: Get a user by username
      operationId: getUserByUsername
      tags:
        - User
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

  /users/role:
    get:
      summary: Get current user's role
      operationId: getMyRole
      tags:
        - User
      responses:
        '200':
          description: Successfully retrieved current user's role
          content:
            application/json:
              schema:
                type: string
                enum: [ROLE_ADMIN, ROLE_USER, ROLE_WORKER, ROLE_MANAGER]
                example: ROLE_USER
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - BearerAuth: []

  /users/myid:
    get:
      summary: Get current user's id
      operationId: getMyId
      tags:
        - User

      responses:
        '200':
          description: Successfully retrieved current user's id
          content:
            application/json:
              schema:
                type: string

        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - BearerAuth: [ ]


  /users/myManager:
    get:
      summary: Get current user's manager
      operationId: getMyManager
      tags:
        - User

      responses:
        '200':
          description: Successfully retrieved current user's manager
          content:
            application/json:
              schema:
                type: string

        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - BearerAuth: [ ]


  /auth/login:
    post:
      summary: Authenticate user
      operationId: login
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Invalid credentials
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Missing or invalid authentication

  schemas:
    UserRequest:
      type: object
      required:
        - username
        - password
        - firstName
        - lastName
        - department
        - role
      properties:
        username:
          type: string
        password:
          type: string
          format: password
        firstName:
          type: string
        lastName:
          type: string
        department:
          type: string
        managerUsername:
          type: string
        role:
          type: string
          enum: [ ROLE_ADMIN, ROLE_USER, ROLE_WAREHOUSE, ROLE_MANAGER ]

    UserPatchRequest:
      type: object
      properties:
        password:
          type: string
          format: password
        firstName:
          type: string
        lastName:
          type: string
        department:
          type: string
        managerUsername:
          type: string
        role:
          type: string

    UserResponse:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        department:
          type: string
        managerUsername:
          type: string
        role:
          type: string
        created:
          type: string
        lastModified:
          type: string
        modifiedBy:
          type: string